catalog-service/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
